refactor: create route for event #19 and #29
