+++ src/js/components/event/index.js
